requirements.txt
streamlit
pandas
matplotlib
openai
langchain
psycopg2
python-dotenv